cravecrafter_backend/api/serializers.py
from rest_framework import serializers
from .models import Restaurant, User 
from rest_framework.validators import UniqueValidator
from django.contrib.auth.models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'email', 'is_active', 'created', 'updated']
        read_only_field = ['is_active', 'created', 'updated']

class RestaurantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Restaurant
        fields = ('id', 'name', 'img_url')

