requirements.txt
asgiref==3.7.2
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
alembic==1.10.3
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
boto==2.49.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
django-markwhat==1.6.2
djangorestframework-simplejwt==5.3.0
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
filelock==3.13.1
Flask==2.2.3
flask-form==1.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
flask-optional-routes==1.4.1
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==5.7.1
Mako==1.2.4
markdown2==2.4.10
MarkupSafe==2.1.2
packaging==23.0
pbr==6.0.0
pickleshare==0.7.5
platformdirs==4.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
sqlparse==0.4.4
stevedore==5.1.0
tornado==6.3.3
traitlets==5.9.0
urllib3==1.26.15
virtualenv==20.24.7
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
